Split processor player queries
